Remove irritating clicking sound
We now run each note on its own angle value, which avoid the annoying
clicks that happened before. To do this, we now have an Oscillator
for each note, that tracks all the state needed to do audio generation.

We also now generate things in batches of buffer size, instead of one
by one. This is hopefully faster(?).