Pass along keypresses into iced_baseview.
This change allows iced_baseview to recieve keypresses from the GUI.
This is needed because `iced_audio` widgets react to certain keypresses
(for example, when holding control, they can be fine tuned), but the host
captures all keypress events, so iced doesn't see them by default. To
fix this, we pass along the keypress into iced_baseview via the Handle
it returns. Note that this requires modifying iced_baseview to allow for
this, so we now depend on our own fork of iced_baseview.