Fix knob flicker bug that happens with Ableton.
It seems that in Ableton, it will end up sending ghost/echoed `set_parameter`
calls after the GUI changes a parameter. This results in weird flickery
behavior with the knobs as the GUI and Ableton race eachother to alter the
parameters. To fix this, we now check if Ableton is trying to set a parameter
to the same value as it already is, and if it is, we just ignore it.

This shouldn't affect other VST hosts that do not send such ghost messages.