Convert notifier to broadcaster.
Previously, the notifier that was used for ForceRedraws was causing us
to update every knob every time a redraw request was sent. This is very
bad if you had many knobs since there's like 100 of them. Instead, we
now send a ForceRedraw that targets the specific knob that was changed.