Generate `set_by_preset` via a macro. Add preset options.
The function `set_by_preset` is now generated via a macro. This adds
two new columns in the parameter table: easer_field and preset_field.

easer_field is the field that is read from EASER when converting into
a normalizd float, while preset_field is the field read from to get the
preset data.