Clamp lerp and inv_lerp to [0.0, 1.0] range. This fixes an envelope bug.
Before, lerp and inv_lerp did not clamp their input/outputs to the 0.0-1.0
range. This could result in problems when t values above 1.0 are inputted,
which happens during the release phase of envelopes if one oscillator happens
to have a long release and the other has a short release (the short release
oscillator will have a period where its t values are above 1.0).

This commit fixes this problem by adding a clamp on lerp and inv_lerp so
that envelopes do not go past their min value.